questao02.py corrigida

+++ questao02.py
@@ -1,4 +1,4 @@
-idade = int(input('Qual é a sua idade?'))
+idade = float(input('Qual é a sua idade?'))
 
 if idade >= 18:
 	print('Maior de idade')
